requirements.txt
Flask
onnxruntime
Pillow
numpy<2.0